+++ lib/plugins/jsx/slots/slots.d.ts
@@ -1,5 +1,0 @@
-import { JSXPluginElement } from '@innet/jsx';
-export interface SlotsProps {
-    from: any;
-}
-export declare function slots({ props: { from }, children }: JSXPluginElement<SlotsProps>, handler: any): any;