+++ src/plugins/jsx/delay/delay.ts
@@ -1,10 +1,9 @@
-import { JSXPluginElement, useHandler } from '@innet/jsx'
+import { Context, JSXPluginElement, useHandler } from '@innet/jsx'
 import innet, { Handler } from 'innet'
 import { onDestroy, scope, State, Watch } from 'watch-state'
 
 import { getComment, Ref } from '../../../utils'
 import { REMOVE_DELAY } from '../../../utils/dom/constants'
-import { Context } from '../context'
 
 export interface DelayProps {
   show?: number